Restructure auth to extract collection of username
