+++ controllers/authentication.go
@@ -1,6 +1,8 @@
 package controllers
 
 import (
+	"net/http"
+
 	"github.com/golang-jwt/jwt"
 	"github.com/hay-i/gooal/auth"
 	"github.com/hay-i/gooal/views"
@@ -13,7 +15,10 @@ func JwtAuthenticationMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
 		if err != nil {
 			views.AddFlash(c, "You must be logged in to access that page", views.FlashError)
 
-			return redirect(c, "/login")
+			// TODO: http.StatusSeeOther
+			// Old note: I wanted to send in http.StatusCreated, but it seems that the redirect doesn't work with that status code
+			// See: https://github.com/labstack/echo/issues/229#issuecomment-1518502318
+			return c.Redirect(http.StatusSeeOther, "/login")
 		}
 
 		tokenString := cookie.Value
@@ -26,7 +31,7 @@ func JwtAuthenticationMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
 		if err != nil || !token.Valid {
 			views.AddFlash(c, "Invalid or expired token", views.FlashError)
 
-			return redirect(c, "/login")
+			return c.Redirect(http.StatusSeeOther, "/login")
 		}
 
 		return next(c)
