+++ controllers/templates.go
@@ -9,7 +9,6 @@ import (
 	"github.com/hay-i/gooal/db"
 	"github.com/hay-i/gooal/formparser"
 	"github.com/hay-i/gooal/models"
-	"github.com/hay-i/gooal/views"
 
 	"github.com/labstack/echo/v4"
 )
@@ -23,25 +22,18 @@ func Build() echo.HandlerFunc {
 		goal := c.QueryParam("goal")
 		focus := c.QueryParam("focus")
 
-		// TODO: Extract all this
-		cookie, err := c.Cookie("token")
+		tokenString, err := auth.GetTokenFromCookie(c)
 		if err != nil {
-			views.AddFlash(c, "You must be logged in to access that page", views.FlashError)
-
-			return redirect(c, "/login")
+			return auth.HandleInvalidToken(c, "You must be logged in to access that page")
 		}
-
-		tokenString := cookie.Value
 
 		parsedToken, err := auth.ParseToken(tokenString)
-
 		if err != nil {
-			views.AddFlash(c, "Invalid or expired token", views.FlashError)
-
-			return redirect(c, "/login")
+			return auth.HandleInvalidToken(c, "Invalid or expired token")
 		}
 
-		component := components.Build(goal, focus, parsedToken["sub"].(string))
+		username := auth.TokenToUsername(parsedToken)
+		component := components.Build(goal, focus, username)
 
 		return renderNoBase(c, component)
 	}
@@ -115,25 +107,18 @@ func Complete(database *mongo.Database) echo.HandlerFunc {
 			return renderNoBase(c, components.Complete(template, questionViews))
 		}
 
-		// TODO: Extract all this
-		cookie, err := c.Cookie("token")
+		tokenString, err := auth.GetTokenFromCookie(c)
 		if err != nil {
-			views.AddFlash(c, "You must be logged in to access that page", views.FlashError)
-
-			return redirect(c, "/login")
+			return auth.HandleInvalidToken(c, "You must be logged in to access that page")
 		}
-
-		tokenString := cookie.Value
 
 		parsedToken, err := auth.ParseToken(tokenString)
-
 		if err != nil {
-			views.AddFlash(c, "Invalid or expired token", views.FlashError)
-
-			return redirect(c, "/login")
+			return auth.HandleInvalidToken(c, "Invalid or expired token")
 		}
 
-		db.SaveAnswer(database, models.Answer{}.FromForm(template.ID, parsedToken["sub"].(string), questionViews))
+		username := auth.TokenToUsername(parsedToken)
+		db.SaveAnswer(database, models.Answer{}.FromForm(template.ID, username, questionViews))
 
 		return renderNoBase(c, components.Save("Response to template"))
 	}
